requirements.txt
alembic==1.4.1
altair==4.1.0
argcomplete
argon2-cffi
astor==0.8.1
async-generator==1.10
attrs
backcall
backports.functools-lru-cache
backports.zoneinfo==0.2.1
base58==2.1.1
bleach
blinker==1.4
blis==0.7.5
bokeh==2.4.1
Boruta==0.3
cachetools==4.2.4
catalogue==1.0.0
certifi==2021.10.8
cffi
charset-normalizer==2.0.7
click==7.1.2
cloudpickle==2.0.0
colorama
colorcet==2.0.6
colorlover==0.3.0
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.6
Cython==0.29.14
databricks-cli==0.16.2
debugpy
decorator
defusedxml
docker==5.0.3
entrypoints
et-xmlfile==1.1.0
Flask==2.0.2
funcy==1.16
future==0.18.2
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.24
greenlet==1.1.2
holoviews==1.14.6
htmlmin==0.1.12
hvplot==0.7.3
idna==3.3
ImageHash==4.2.1
imbalanced-learn==0.7.0
importlib-metadata
importlib-resources==5.4.0
ipykernel==6.5.0
ipython
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi
Jinja2
joblib==1.0.1
jsonschema==4.2.0
jupyter-client
jupyter-core
jupyterlab-pygments
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
kmodes==0.11.1
lightgbm==3.3.1
llvmlite==0.36.0
Mako==1.1.5
Markdown==3.3.4
MarkupSafe
matplotlib==3.4.3
matplotlib-inline
missingno==0.5.0
mistune
mlflow==1.21.0
mlxtend==0.19.0
mols2grid
multimethod==1.6
murmurhash==1.0.6
nbclient
nbconvert
nbformat
nest-asyncio
networkx==2.6.3
nltk==3.6.5
notebook
numba==0.53.1
numexpr==2.7.3
numpy==1.19.5
openpyxl==3.0.9
packaging==21.2
pandas==1.3.4
pandas-profiling==3.1.0
pandocfilters
panel==0.12.4
param==1.12.0
parso
patsy==0.5.2
phik==0.12.0
pickleshare
Pillow==8.4.0
plac==1.1.3
plotly==5.3.1
plotly-express==0.4.1
preshed==3.0.6
prometheus-client 
prometheus-flask-exporter==0.18.5
prompt-toolkit
protobuf==3.19.1
pyarrow==6.0.0
pycaret==2.3.4
pycparser
pyct==0.4.8
pydantic==1.8.2
pydeck==0.7.1
Pygments
pyLDAvis==3.2.2
pynndescent==0.5.5
pyod==0.9.5
pyparsing==2.4.7
pyrsistent
python-dateutil
python-editor==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.1.0
PyWavelets==1.1.1
pywinpty
PyYAML==6.0
pyzmq
querystring-parser==1.2.4
regex==2021.11.2
requests==2.26.0
rdkit-pypi
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.2
Send2Trash
six
smart-open==5.2.1
smmap==5.0.0
spacy==2.3.7
SQLAlchemy==1.4.26
sqlparse==0.4.2
srsly==1.0.5
statsmodels==0.13.0
streamlit==1.1.0
streamlit-lottie==0.0.2
streamlit-pandas-profiling==0.1.2
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
terminado
testpath
textblob==0.17.1
thinc==7.4.5
threadpoolctl==3.0.0
toml==0.10.2
toolz==0.11.1
tornado
tqdm==4.62.3
traitlets
typing-extensions
tzdata==2021.5
tzlocal==4.1
umap-learn==0.5.2
urllib3==1.26.7
validators==0.18.2
visions==0.7.4
waitress==2.0.0
wasabi==0.8.2
watchdog==2.1.6
wcwidth
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wordcloud==1.8.1
yellowbrick==1.3.post1
zipp
